+++ src/langchain_google_bigtable/__init__.py
@@ -13,13 +13,25 @@
 # limitations under the License.
 
 
-from .chat_message_history import BigtableChatMessageHistory, create_chat_history_table
-from .loader import BigtableLoader, BigtableSaver, Encoding, MetadataMapping
+from .chat_message_history import (
+    BigtableChatMessageHistory,
+    create_chat_history_table,
+    init_chat_history_table,
+)
+from .loader import (
+    BigtableLoader,
+    BigtableSaver,
+    Encoding,
+    MetadataMapping,
+    init_document_table,
+)
 from .version import __version__
 
 __all__ = [
     "BigtableChatMessageHistory",
     "create_chat_history_table",
+    "init_document_table",
+    "init_chat_history_table",
     "BigtableLoader",
     "BigtableSaver",
     "MetadataMapping",